single updates

+++ proteins/data/parse.py
@@ -1,26 +1,72 @@
 import subprocess
 from os import path
 import pandas as pd
+from esm.constants import proteinseq_toks
 
-root_dir = "./data_files"
+data_dir = "./data_files"
 
-def get_tdc_epitope(name, split=False, file_dir=root_dir):
+def sanitize_sequence(seq):
+    return ''.join(c if c in proteinseq_toks['toks'] else 'X' for c in seq)
+
+def get_tdc_epitope(name, file_dir=data_dir):
     from tdc.single_pred import Epitope
-    data = Epitope(name=name,  path=file_dir)
-    return data.get_split() if split else data.get_data()
+    mapping = {'Antigen_ID': 'ID', 'Antigen': 'X'}
+    data = Epitope(name=name,  path=file_dir).get_data().rename(columns=mapping)
+    data['ID'] = data['ID'].str.replace(" ", "")
+    return data
 
-def get_tdc_ppi(name, split=False, neg_frac=1, file_dir=root_dir):
+def get_tdc_ppi(name, split=False, neg_frac=1, file_dir=data_dir):
     from tdc.multi_pred import PPI
     data = PPI(name=name, path=file_dir).neg_sample(frac=neg_frac)
     return data.get_split() if split else data.get_data()
 
-def df_save(data, name, file_dir=root_dir):
-    file_path = path.join(file_dir, name+'.parquet')
-    data.to_parquet(file_path, engine='pyarrow')
+def make_sequence_fasta(
+    sequences,
+    ids,
+    file_dir=data_dir,
+    force=False
+):
+    fasta_path = path.join(file_dir, "sequences.fasta")
+    if force or not path.exists(fasta_path):
+        from Bio.Seq import Seq
+        from Bio.SeqRecord import SeqRecord
+        from Bio import SeqIO
+        records = []
+        for seq, sid in zip(sequences, ids):
+            seq = sanitize_sequence(seq)
+            records.append(SeqRecord(Seq(seq), id=sid, description=""))
+        SeqIO.write(records, fasta_path, "fasta")
+        print(f'Wrote sequences to {fasta_path}')
+    else:
+        print("Fasta file already exists; set force=True to overwrite")
+    return fasta_path
 
-def df_load(name, file_dir=root_dir):
-    file_path = path.join(file_dir, name+'.parquet')
+
+def df_save(data, name='dataframe', file_dir=data_dir, force=False):
+    file_path = path.join(file_dir, f'{name}.parquet')
+    if force or not path.exists(file_path):
+        data.to_parquet(file_path, engine='pyarrow')
+        print(f'Saved dataframe to {file_path}')
+
+
+def df_load(data_name, file_dir=data_dir):
+    file_path = path.join(file_dir, f'{data_name}.parquet')
     return pd.read_parquet(file_path, engine='pyarrow')
+
+def cluster_fasta(fasta_path, cluster_coef=0.5, force=False):
+    base = path.dirname(fasta_path)
+    output_prefix = path.join(base, "clustered_sequences")
+    clstr_file = output_prefix + ".clstr"
+    if force or not path.exists(clstr_file):
+        if not force:
+            print("Clustering file not found, generating new")
+        subprocess.run([
+            "cd-hit", "-i", fasta_path, "-o", output_prefix,
+            "-c", str(cluster_coef), "-n", "2", "-M", "16000", "-T", "8"
+        ], check=True)
+    else:
+        print("Clustering file found, parsing in data")
+    return clstr_file
 
 def parse_cd_hit_clstr(clstr_file, seq_ids_order):
     cluster_map = {}
@@ -34,42 +80,11 @@ def parse_cd_hit_clstr(clstr_file, seq_ids_order):
                 cluster_map[seq_id] = cluster_id
     return cluster_map
 
-def cluster_sequences(data, data_id, cluster_coef=0.4, force=False, sequence_col='Antigen', id_col='Antigen_ID'):
-    from Bio import SeqIO
-    from Bio.Seq import Seq
-    from Bio.SeqRecord import SeqRecord
-    data=data.copy()
-    data[id_col] = data[id_col].str.replace(' ', '')
-    fasta_path = f"./raw_data_files/{data_id}.fasta"
-    if force or not path.exists(fasta_path):
-        records = []
-        for idx, row in data.iterrows():
-            seq = row[sequence_col]
-            antigen_id = row.get(id_col, f"antigen_{idx}")
-            records.append(SeqRecord(Seq(seq), id=antigen_id, description=""))
-        SeqIO.write(records, fasta_path, "fasta")
-    output_fasta = f"./raw_data_files/{data_id}_clustered.fasta"
-    clstr_file = output_fasta + ".clstr"
-    if force or not path.exists(clstr_file):
-        if not force:
-            print("Clustering file not found, generating new")
-        subprocess.run([
-            "cd-hit", "-i", fasta_path, "-o", output_fasta,
-            "-c", str(cluster_coef), "-n", "2", "-M", "16000", "-T", "8"
-        ], check=True)
-    else:
-        print("Clustering file found, parsing in data")
-
-    antigen_ids = [row.get(id_col, f"antigen_{i}") for i, row in data.iterrows()]
-    cluster_map = parse_cd_hit_clstr(clstr_file, set(antigen_ids))
+def add_clusters_to_df(df, clstr_path):
+    data = df.copy()
+    ids = df['ID'].replace(" ", "")
+    cluster_map = parse_cd_hit_clstr(clstr_path, set(ids))
     if not cluster_map:
-        if force:
-            raise Exception("Error while clustering")
-        else:
-            print("No results found in old cluster file, generating new")
-            return cluster_sequences(data, data_id, cluster_coef=cluster_coef, force=True,
-                              sequence_col=sequence_col, id_col=id_col)
-
-    data['cluster_id'] = data[id_col].map(cluster_map)
+        raise Exception("Error while loading cluster file")
+    data['cluster_id'] = ids.map(cluster_map)
     return data
-
